[tools/onert_train] Separate DataLoader class
ONE-DCO-1.0-Signed-off-by: Jiyoung Yun <jy910.yun@samsung.com>
